pyproject.toml
[project]
name = "sabot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "shlyashl", email = "ay-ya@ya.ru"},
]
dependencies = ["aiogram>=3.15.0", "requests>=2.32.3", "python-dotenv>=1.0.1"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "python src/sabot/bot.py"
